it builds

+++ DataLayer/ItemsDBO.h
@@ -16,11 +16,11 @@ INT32 SerializeItem(Item* i, MemoryStream * outData);
 INT32 DeserializeItem(Item* i, MemoryStream * data);
 
 Item *CreateItem(Item * existing, UID playerId, sql::Connection * sql);
-Item *CreateItemFromTemplate(UINT64 iTemplateIndex, UID playerId, sql::Connection * sql);
+Item *CreateItemFromTemplate(UINT32 iTemplateIndex, UID playerId, sql::Connection * sql);
 Item *CreateItem(UINT64 itemId, UID playerId, sql::Connection * sql);
 
 Item* CreateNonDbItem(Item * existing);
-Item* CreateNonDbItemFromTemplate(UINT64 iTemplateIndex, UINT32 stackCount);
+Item* CreateNonDbItemFromTemplate(UINT32 iTemplateIndex, UINT32 stackCount);
 Item* CreateNonDbItem(UINT64 itemId, UINT32 stack);
 
 INT32 UpdateItem(Item * i, sql::Connection * conn);
